changes in RUT
