+++ packages/models-library/src/models_library/api_schemas_resource_usage_tracker/credit_transactions.py
@@ -1,5 +1,10 @@
+from datetime import datetime
 from decimal import Decimal
 
+from models_library.products import ProductName
+from models_library.resource_tracker import CreditTransactionId
+from models_library.users import UserID
+from models_library.wallets import WalletID
 from pydantic import BaseModel, validator
 
 from ..wallets import WalletID
@@ -13,3 +18,20 @@ class WalletTotalCredits(BaseModel):
     @classmethod
     def ensure_rounded(cls, v):
         return round(v, 2)
+
+
+class CreditTransactionCreateBody(BaseModel):
+    product_name: ProductName
+    wallet_id: WalletID
+    wallet_name: str
+    user_id: UserID
+    user_email: str
+    osparc_credits: Decimal
+    payment_transaction_id: str
+    created_at: datetime
+
+
+class CreditTransactionCreated(BaseModel):
+    """Response Create Credit Transaction V1 Credit Transactions Post"""
+
+    credit_transaction_id: CreditTransactionId
